Update 2020-02-10-ACCN.md
Joel hasn't directly contacted me, and I in no way feel that I have violated my duty to fairly and accurate cite the work I've built upon; however, Joel expressed concerns on twitter re GlyphNet being seen as the origin of CCNs when it in fact came after. I previously cited both Dimensions of Dialogue and GlyphNet, but am adding that Joel's Dimensions of Dialogue is the seminal work in this small area of study.